# simple-erp/erp-api/openapi.yaml
...
openapi: 3.0.3
info:
  title: REST API for study ERP system
  contact:
    name: Integrity Vision LLC
    url: https://integrity.com.ua
    email: vivanchenko@integrity.com.ua
  version: "1.0"
servers:
- url: /erp
paths:
...
  /api/rest/material:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaterialDto'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaterialDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaterialDto'
  /api/rest/material/batch:
    get:
      parameters:
      - name: material
        in: query
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatchOfMaterialDto'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchOfMaterialDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchOfMaterialDto'
  /api/rest/material/batch/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: Return batch data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchOfMaterialDto'
        "404":
          description: Batch not exist
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "204":
          description: No Content
    patch:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchWriteOffRequestDto'
      responses:
        "204":
          description: No Content
  /api/rest/material/consumption-rate:
    get:
      parameters:
      - name: material
        in: query
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsumptionRateDto'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumptionRateDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsumptionRateDto'
  /api/rest/material/consumption-rate/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: Return consumption rate data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsumptionRateDto'
        "404":
          description: Consumption rate not exist
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumptionRateDto'
      responses:
        "204":
          description: No Content
  /api/rest/org-structure/employee:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeDto'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
  /api/rest/org-structure/employee/login/{login}:
    get:
      parameters:
      - name: login
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Return employee data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
        "404":
          description: Employee not exist
  /api/rest/org-structure/employee/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: Return employee data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDto'
        "404":
          description: Employee not exist
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeDto'
      responses:
        "204":
          description: No Content
  /api/rest/org-structure/position:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PositionDto'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionDto'
  /api/rest/org-structure/position/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: Return position data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionDto'
        "404":
          description: Position not exist
  /api/rest/org-structure/position/{id}/supply-rate:
    patch:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyRateDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionDto'
  /api/rest/org-structure/position/{id}/supply-rate/{rateId}:
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      - name: rateId
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionDto'
  /api/rest/org-structure/unit:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrgUnitDto'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgUnitDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgUnitDto'
  /api/rest/org-structure/unit/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: Return unit data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgUnitDto'
        "404":
          description: Unit not exist
  /api/rest/org-structure/vacation:
    get:
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VacationDto'
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VacationDto'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VacationDto'
  /api/rest/org-structure/vacation/{id}:
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "204":
          description: No Content
...
components:
  schemas:
    BatchOfMaterialDto:
      required:
      - material
      - purchaseDate
      - purchaseAmount
      - price
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        material:
          allOf:
          - $ref: '#/components/schemas/MaterialDto'
          - nullable: false
        purchaseDate:
          format: date
          type: string
          xml:
            name: purchase-date
          nullable: false
        purchaseAmount:
          type: number
          xml:
            name: purchase-amount
          nullable: false
        price:
          type: number
          nullable: false
        remainingAmount:
          type: number
          xml:
            name: remaining-amount
      xml:
        name: batch-of-material
    BatchWriteOffRequestDto:
      required:
      - batchId
      - amount
      type: object
      properties:
        batchId:
          format: int64
          type: integer
          xml:
            name: batch-id
            attribute: true
          nullable: false
        amount:
          pattern: \S
          type: string
          xml:
            attribute: true
          nullable: false
      xml:
        name: write-off-request
    ConsumptionRateDto:
      required:
      - title
      - material
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        title:
          pattern: \S
          type: string
          nullable: false
        material:
          allOf:
          - $ref: '#/components/schemas/MaterialDto'
          - nullable: false
        durationUnit:
          type: object
          xml:
            name: duration-unit
        durationAmount:
          format: int64
          type: integer
          xml:
            name: duration-amount
        amount:
          type: number
      xml:
        name: consumption-rate
    EmployeeDto:
      required:
      - login
      - firstName
      - lastName
      - birthDate
      - hireDate
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        login:
          pattern: \S
          type: string
          nullable: false
        firstName:
          pattern: \S
          type: string
          xml:
            name: first-name
          nullable: false
        lastName:
          pattern: \S
          type: string
          xml:
            name: last-name
          nullable: false
        birthDate:
          format: date
          type: string
          xml:
            name: birth-date
          nullable: false
        hireDate:
          format: date
          type: string
          xml:
            name: hire-date
          nullable: false
        fireDate:
          format: date
          type: string
          xml:
            name: fire-date
        position:
          format: int64
          type: integer
      xml:
        name: employee
...
    MaterialDto:
      required:
      - title
      - measureUnit
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        title:
          pattern: \S
          type: string
          nullable: false
        measureUnit:
          allOf:
          - $ref: '#/components/schemas/MeasureUnit'
          - xml:
              name: measure-unit
            nullable: false
      xml:
        name: material
    MeasureUnit:
      enum:
      - METER
      - SQUARE_METER
      - KILOGRAMME
      - LITER
      - PCS
      - PKG
      type: string
    OrgUnitDto:
      required:
      - title
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        title:
          pattern: \S
          type: string
          nullable: false
        ascendant:
          format: int64
          type: integer
          xml:
            name: ascendant-unit
        descendants:
          type: array
          items:
            format: int64
            type: integer
            xml:
              name: unit
          xml:
            name: descendant-units
            wrapped: true
        positions:
          type: array
          items:
            format: int64
            type: integer
            xml:
              name: position
          xml:
            wrapped: true
      xml:
        name: unit
    PositionDto:
      required:
      - title
      - index
      - orgUnit
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        title:
          pattern: \S
          type: string
          nullable: false
        index:
          format: int32
          type: integer
          nullable: false
        orgUnit:
          format: int64
          type: integer
          xml:
            name: unit
          nullable: false
        employee:
          format: int64
          type: integer
        supplyRates:
          type: array
          items:
            format: int64
            type: integer
            xml:
              name: rate
          xml:
            name: supply-rates
            wrapped: true
      xml:
        name: position
    SupplyRateDto:
      required:
      - title
      - equipmentType
      - amount
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        title:
          pattern: \S
          type: string
          nullable: false
        equipmentType:
          format: int64
          type: integer
          xml:
            name: equipment-type
          nullable: false
        amount:
          format: int32
          type: integer
          nullable: false
      xml:
        name: supply-rate
...
    VacationDto:
      required:
      - employee
      - startDate
      - duration
      - type
      type: object
      properties:
        id:
          format: int64
          type: integer
          xml:
            attribute: true
        employee:
          format: int64
          type: integer
          nullable: false
        startDate:
          format: date
          type: string
          xml:
            name: start-date
          nullable: false
        duration:
          format: int32
          type: integer
          nullable: false
        type:
          allOf:
          - $ref: '#/components/schemas/VacationType'
          - nullable: false
      xml:
        name: vacation
    VacationType:
      enum:
      - ANNUALLY
      - ADDITIONAL
      - AT_OWN_EXPENSE
      - FOR_FAMILY_REASONS
      type: string